refactor: namings, methods
