+++ packages/core-snapshots-cli/lib/commands/slice.js
@@ -1,18 +1,16 @@
 'use strict'
 const zlib = require('zlib')
-const init = require('../init')
 const StreamValues = require('stream-json/streamers/StreamValues')
 const fs = require('fs-extra')
 const cliProgress = require('cli-progress')
-const container = require('@arkecosystem/core-container')
-const logger = container.resolvePlugin('logger')
+const gzip = require('../utils/gzip')
 
 module.exports = async (options) => {
   const storageLocation = `${process.env.ARK_PATH_DATA}/snapshots/${process.env.ARK_NETWORK_NAME}`
   const snapshotHeight = options.filename ? parseInt(options.filename.split('.')[1]) : 0
 
   await fs.ensureFile(`${storageLocation}/slice.dat`)
-  const rollbackStream = fs.createWriteStream(`${storageLocation}/slice.dat`)
+  const sliceStrem = fs.createWriteStream(`${storageLocation}/slice.dat`)
   const sourceStream = fs.createReadStream(`${storageLocation}/${options.filename}`)
 
   if (options.end === 0) {
@@ -27,32 +25,20 @@ module.exports = async (options) => {
 
     pipeline.on('data', (data) => {
       if (data.value.height >= options.start && data.value.height <= options.end) {
-        rollbackStream.write(`{"height":${data.value.height}, "blockBuffer":"${data.value.blockBuffer}"}`)
+        sliceStrem.write(`{"height":${data.value.height}, "blockBuffer":"${data.value.blockBuffer}"}`)
 
         progressBbar.update(data.value.height)
       }
 
       if (data.value.height > options.end) {
         sourceStream.close()
-        rollbackStream.close()
+        sliceStrem.close()
 
         progressBbar.stop()
       }
   })
 
-  const __gzip = (sourceFileName, height) => {
-    fs.createReadStream(`${storageLocation}/${sourceFileName}`)
-      .pipe(zlib.createGzip())
-      .pipe(fs.createWriteStream(`${storageLocation}/snapshot.${height}.gz`))
-      .on('finish', async () => {
-        fs.unlinkSync(`${storageLocation}/${sourceFileName}`)
-        logger.info(`New snapshot was succesfully created. File: [snapshot.${height}.gz]`)
-
-        await init.tearDown()
-      })
-  }
-
   sourceStream.on('close', async () => {
-    __gzip('slice.dat', options.end)
+    gzip('slice.dat', options.end)
   })
 }
